moved ErrorType in views
