+++ answerking_app/tests/test_items.py
@@ -2,7 +2,8 @@ from django.test import TestCase, Client
 from answerking_app.models.models import Item
 
 from django.db.models.query import QuerySet
-from API_types import NewItemType, ItemType, ItemIDType, ErrorMessage
+from answerking_app.views.ErrorType import ErrorMessage
+from API_types import NewItemType, ItemType, ItemIDType
 
 client = Client()
