+++ answerking_app/views/ErrorType.py
@@ -1,0 +1,10 @@
+from typing import TypedDict
+
+
+class ErrorMessageContent(TypedDict):
+    message: str
+    details: str
+
+
+class ErrorMessage(TypedDict):
+    error: ErrorMessageContent
